+++ src/main/java/com/web/service/PostService.java
@@ -1,19 +1,18 @@
 package com.web.service;
 
 import java.io.IOException;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.web.data.Group;
 import com.web.data.Post;
 import com.web.data.User;
-import com.web.data.Group;
 import com.web.data.dto.PostDto;
 import com.web.repository.PostRepo;
-import com.web.utils.DateUtil;
 
 @Service
 public class PostService {
@@ -46,25 +45,31 @@ public class PostService {
 		return post;
 	}
 	
-	public void updatePost(Post post, String text, String tags, MultipartFile file) throws IllegalStateException, IOException {
-		post.setPostText(text);
-		post.setTags(tags);
-		post.setPostCreationDate(LocalDateTime.now());
-		post.setFilename(fileService.uploadFile(file, UploadType.POST));
-		postRepo.save(post);
+	public void updatePost(User currentUser, Post post, String text, String tags, MultipartFile file) throws IllegalStateException, IOException {
+		if(userIsAuthorOrAdmin(currentUser, post)) {
+			post.setPostText(text);
+			post.setTags(tags);
+			post.setPostCreationDate(LocalDateTime.now());
+			post.setFilename(fileService.uploadFile(file, UploadType.POST));
+			postRepo.save(post);
+		}
 	}
 
-	public void deletePost(Post post, User currentUser) {
-		if(postRepo.findCountByRepostAndAuthor(currentUser, post.getRepost()) == 1)  {
-			decrementAndSaveRepostsCount(post.getRepost());
+	public void deletePost(User currentUser, Post post) {
+		if(userIsAuthorOrAdmin(currentUser, post)) {
+			if(postRepo.findCountByRepostAndAuthor(currentUser, post.getRepost()) == 1)  {
+				decrementAndSaveRepostsCount(post.getRepost());
+			}
+			postRepo.delete(post);
 		}
-		postRepo.delete(post);
 	}
 	
-	public void removeRepost(Post post) {
-		decrementAndSaveRepostsCount(post.getRepost());
-		post.setRepost(null);
-		postRepo.save(post);
+	public void removeRepost(User currentUser, Post post) {
+		if(userIsAuthorOrAdmin(currentUser, post)) {
+			decrementAndSaveRepostsCount(post.getRepost());
+			post.setRepost(null);
+			postRepo.save(post);
+		}
 	}
 	
 	public void addRepost(User currentUser, Post post, String repostText, String repostTags) throws IllegalStateException, IOException {
@@ -86,6 +91,17 @@ public class PostService {
 		postRepo.save(decrimentedPost);
 	}
 	
+	public void like(User currentUser, User user, Post post) {
+		if(user.equals(currentUser)) {
+			Set<User> like = post.getPostLikes();
+			if(like.contains(user)) {
+				like.remove(user);
+			} else {
+				like.add(user);
+			}
+		}
+	}
+	
 	public PostDto findOnePost(User currentUser, Post post) {
 		return postRepo.findOnePost(currentUser, post.getId());
 	}
@@ -98,6 +114,11 @@ public class PostService {
 			searchResult = postRepo.findAll(currentUser);
 		}		
 		return searchResult;
+	}
+	
+	public boolean userIsAuthorOrAdmin(User currentUser, Post post) {
+		return post.getPostAuthor() != null && post.getPostAuthor().equals(currentUser)
+			   || (post.getPostGroup() != null && post.getPostGroup().getGroupAdmins().contains(currentUser));
 	}
 	
 	public Iterable<PostDto> findPostsByUser(User currentUser, User user) {
